🧹rename: parser2.ts -> parser.ts
